requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
importlib_metadata==7.1.0
Markdown==3.6
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
zipp==3.18.1
